requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
PyJWT==2.8.0
python-dotenv==1.0.0
redis==5.0.1
pandas==2.1.4
openpyxl==3.1.2
APScheduler==3.10.4
email-validator==2.1.0
bcrypt==4.1.2
marshmallow==3.20.1
Flask-Limiter==3.5.0
SQLAlchemy==2.0.25
flask-swagger-ui==4.11.1
requests==2.31.0
bleach==6.1.0
python-dateutil==2.8.2
gunicorn==21.2.0
gevent==23.9.1
psutil==5.9.8
